fix: use float-precision literals and math functions in 11 kernels
Replace double-precision literals and functions with float equivalents
to avoid unnecessary float-to-double-to-float promotion in single-precision kernels:

- 1.0 / scalar to 1.0f / scalar (9 kernels)
- pow() to powf() in volk_32f_stddev_and_mean_32f_x2
- tan() to tanf() in volk_32f_tan_32f tail loops
- conj() to lv_conj() in volk_32fc_x2_multiply_conjugate_32fc NEON tail loop

+++ kernels/volk/volk_32f_stddev_and_mean_32f_x2.h
@@ -104,7 +104,7 @@ static inline void volk_32f_stddev_and_mean_32f_x2_generic(float* stddev,
         SquareSum[1] += r * powf(n_plus_one * Val1 - Sum[1], 2);
     }
 
-    SquareSum[0] += SquareSum[1] + .5f / half_points * pow(Sum[0] - Sum[1], 2);
+    SquareSum[0] += SquareSum[1] + .5f / half_points * powf(Sum[0] - Sum[1], 2);
     Sum[0] += Sum[1];
 
     uint32_t points_done = half_points * 2;
